services/dream-poc/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pillow-heif
numpy
opencv-python-headless
rapidocr-onnxruntime
onnxruntime
pdf2image
pdfplumber
pypdfium2
pypdf
ftfy
tqdm
pytesseract
paddleocr
paddlepaddle
torch
torchvision
ultralytics
python-magic
git+https://github.com/openai/CLIP.git@main
aiofiles
requests
python-dotenv
openai
build123d
scipy
scikit-image
diffusers
transformers
accelerate
